l ajout des repositories

+++ src/main/java/com/stagefinder/dao/TypeAbonnementRepository.java
@@ -3,5 +3,5 @@ package com.stagefinder.dao;
 import com.stagefinder.entities.TypeAbonnement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeAbonnementRepository extends JpaRepository<TypeAbonnement, Integer> {
+public interface TypeAbonnementRepository extends JpaRepository<TypeAbonnement, Long> {
 }
